requirements.txt
python-dotenv
Django>3.2,<4.0
django-storages
gunicorn
psycopg2-binary

requests

